requirements.txt
boto3==1.35.26
google-cloud-storage==2.18.0
azure-storage-blob==12.22.0
PyYAML==6.0.2
Requests==2.32.3
playwright>=1.49.1
